adding signup and adding security to crud operations
adding signup function and adding an authentication layer before updating or deleting user

+++ backend/crud/routes.py
@@ -7,6 +7,8 @@ from databases.validators import (UserCreate, UserUpdate, UserResponse,
                                 DocumentCreate, DocumentUpdate, DocumentResponse)
 from utils.security import hash_password 
 
+
+from auth.auth import get_current_user
 router = APIRouter(prefix=os.getenv("ROOT_ENDPOINT"))
 
 @router.get("/users", response_model=list[UserResponse])
@@ -24,22 +26,13 @@ async def get_user_by_id(user_id: int, db: AbstractDatabase = Depends(get_db)):
     return result
 
 
-@router.post("/users")
-async def create_user(user: UserCreate, db: AbstractDatabase = Depends(get_db)):
-    query = "INSERT INTO users (name, email, password) VALUES ($1, $2, $3)"
-    hashed_pwd = hash_password(user.password)
-    try:
-        await db.execute(query, user.name, user.email, hashed_pwd)
-    except Exception as e:
-
-        if "unique constraint" in str(e).lower():
-            raise HTTPException(status_code=400, detail="Email already registered")
-        raise HTTPException(status_code=500, detail="Database error: {}".format(e))
-    return {"message": "User created successfully"}
-    
-
 @router.put("/users/{user_id}", response_model=UserResponse)
-async def update_user(user_id: int, user: UserUpdate, db: AbstractDatabase = Depends(get_db)):
+async def update_user(user_id: int, user: UserUpdate, db: AbstractDatabase = Depends(get_db), current_user: UserResponse = Depends(get_current_user)):
+    # AUTHORIZATION
+    if current_user.id != user_id:
+        raise HTTPException(status_code=403, detail="Unauthorized")
+    
+    # UPDATE USER
     query = "UPDATE users SET name = $2, email = $3, password = $4 WHERE id = $1 RETURNING *"
     hashed_pwd = hash_password(user.password)
     try:
@@ -52,7 +45,12 @@ async def update_user(user_id: int, user: UserUpdate, db: AbstractDatabase = Dep
     return result
 
 @router.delete("/users/{user_id}")
-async def delete_user(user_id: int, db: AbstractDatabase = Depends(get_db)):
+async def delete_user(user_id: int, db: AbstractDatabase = Depends(get_db), current_user: UserResponse = Depends(get_current_user)):
+    # AUTHORIZATION
+    if current_user.id != user_id:
+        raise HTTPException(status_code=403, detail="Unauthorized")
+    
+    # DELETE USER
     query = "DELETE FROM users WHERE id = $1"
     try:
         await db.execute(query, user_id)
